[Fix] createEventBody를 별도로 분리
