+++ src/renderer/src/features/event/ui/add-event/AddEventForm.mutation.tsx
@@ -1,7 +1,7 @@
 import { useLogin } from '@/shared/hooks/useLogin';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { CalendarEvent } from '@/shared/types/EventType';
-import { formatDateLocal, makeDateTime, toIsoStringWithOffset } from '@/shared/lib/dateFunction';
+import { createEventBody } from '@/features/event/lib/createEventBody';
 
 interface AddEventProp {
     date: Date;
@@ -59,31 +59,4 @@ export function useAddEvent() {
     });
 
     return { addEvent: addEventMutation.mutate };
-}
-
-function createEventBody({ date, start, end, summary, colorId, allDay }: AddEventProp) {
-    if (allDay) {
-        const startDate = formatDateLocal(date);
-        const endDateObj = new Date(date);
-        endDateObj.setDate(endDateObj.getDate() + 1);
-        const endDate = formatDateLocal(endDateObj);
-
-        return {
-            summary,
-            start: { date: startDate },
-            end: { date: endDate },
-            colorId: colorId || '1'
-        };
-    }
-
-    const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-    const startDateTime = makeDateTime(date, start);
-    const endDateTime = makeDateTime(date, end);
-
-    return {
-        summary,
-        start: { dateTime: toIsoStringWithOffset(startDateTime), timeZone },
-        end: { dateTime: toIsoStringWithOffset(endDateTime), timeZone },
-        colorId: colorId || '1'
-    };
 }
